Rename deprecatedSecretParamsMap
The struct name deprecatedSecretParamsMap is misleading, it includes the
"deprecated" prefix hinting that the struct it deprecated, however it's
not.  It also isn't made up of strictly deprecated secret keys (which
are in fact labeled with the "deprecated" prefix).

Given this a private struct, it's simple to just rename it, this change
renames it to simply "secretParamsMap" and adds a comment to the struct
to give at least a hint as to the intent.

Addresses Bug #224

Revert "Extend allowed DataSources"

This reverts commit b9148ef01c47650ce53a44c387b841d56053dec1.